New functionalities and bug fixes
Many new functionalities were added to the package for statistics without imputation, metadata exploration, and ID modification on the fly. ANOVA an T.test code were adjusted. To determine if the missingness in the different groups have a biological meaning, a Glm Binomial categorical test function romicsGlmBinomial() was added. Various transformative and plot functions were adjusted to work with data with missingness. New plots enabling the metadata factor to be put in relation (Sankey Chart) and to plot the count of the different levels of each factor. A single variable plot enabling to show the ttest pvalues for a given variable was also added. Manual for the new functionalities were added. some bug fixes were performed.